Add ability to check for an existing properties file This allows us to pass a file in when running the image.

+++ startup.sh
@@ -10,7 +10,14 @@ pip install -q --upgrade -r requirements.txt
 pip install -q --upgrade -r requirements-optional.txt
 
 # Copy the properties file
-cp robottelo.properties.sample robottelo.properties
+if [ ! -e robottelo.properties ]
+then
+	cp robottelo.properties.sample robottelo.properties
+	sed -i "s/^project.*/project=satellite6/" robottelo.properties
+	sed -i "s/^# [robottelo].*/[robottelo]/" robottelo.properties
+	sed -i "s/^# webdriver.*/webdriver=phantomjs/" robottelo.properties
+fi
+
 # Tweak it
 if [ -z "$UPSTREAM" ]
 then
@@ -18,9 +25,7 @@ then
 fi
 sed -i "s/^hostname.*/hostname=${SERVER_URL}/" robottelo.properties
 sed -i "s/^ssh_key.*/ssh_key=${SSH_KEY}/" robottelo.properties
-sed -i "s/^project.*/project=satellite6/" robottelo.properties
 sed -i 's/^upstream.*/upstream=${UPSTREAM}/' robottelo.properties
-sed -i "s/^# [robottelo].*/[robottelo]/" robottelo.properties
-sed -i "s/^# webdriver.*/webdriver=phantomjs/" robottelo.properties
+
 
 nosetests -v ${TESTS}
